Add voronoi plotting for displaying irregular light profiles
